Move initRepos and initServices del go.mod and go.sum dupes, rename user_repository_impl
