+++ backend/bootstrap/init_repositories.go
@@ -1,0 +1,17 @@
+package bootstrap
+
+import (
+	"database/sql"
+	"mellow/repositories"
+	"mellow/repositories/repoimpl"
+)
+
+type Repositories struct {
+	UserRepository repositories.UserRepository
+}
+
+func InitRepositories(db *sql.DB) *Repositories {
+	return &Repositories{
+		UserRepository: repoimpl.NewUserRepository(db),
+	}
+}
