--- go.mod
@@ -1,5 +1,0 @@
-module mellow
-
-go 1.24.2
-
-require github.com/mattn/go-sqlite3 v1.14.28 // indirect
